services/moderation-worker/requirements.txt
numpy==1.26.4
torch==2.1.1
transformers==4.35.2
onnxruntime==1.16.3
optimum[onnxruntime]==1.15.0
underthesea==6.7.0
langdetect==1.0.9
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
aio-pika==9.3.1
python-dotenv==1.0.0
Pillow==10.0.1
openai-whisper==20231117
librosa==0.10.1
requests==2.31.0
soundfile==0.12.1
easyocr==1.7.1
